# fastapi/models.py
# generated by fastapi-codegen:
#   filename:  ./openapi.yml
#   timestamp: 2024-12-18T21:43:25+00:00

from __future__ import annotations

from typing import List, Optional

from pydantic import BaseModel, Field


class StatusResponse(BaseModel):
    status: str = Field(..., title='Status')


class SyncDataResponse(BaseModel):
    status: str = Field(..., title='Status')


class Datum(BaseModel):
    date: Optional[str] = None
    text: Optional[str] = None
    key_rate: Optional[float] = None
    inflation: Optional[float] = None


class DataResponse(BaseModel):
    data: Optional[List[Datum]] = Field(None, title='Data')


class Model(BaseModel):
    model_id: Optional[str] = None
    description: Optional[str] = None
    fit: Optional[bool] = None


class ModelsResponse(BaseModel):
    models: List[Model] = Field(..., title='Models')


class FitResponse(BaseModel):
    message: str = Field(..., title='Message')


class PredictResponse(BaseModel):
    prediction: float = Field(..., title='Prediction')